requirements.txt
python-barcode==0.15.1
charset-normalizer==3.4.4
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.3
seaborn==0.13.2
six==1.17.0
tzdata==2025.2
